chore: upgrade probot to v9.6.6
