+++ package.json
@@ -14,32 +14,32 @@
     "postinstall": "tsc"
   },
   "dependencies": {
-    "@octokit/rest": "^16.15.0",
-    "@types/ioredis": "^4.0.12",
+    "@octokit/rest": "^16.35.2",
+    "@types/ioredis": "^4.14.2",
     "@types/lru-cache": "^5.1.0",
     "config-yml": "^0.10.2",
     "fs-extra": "^7.0.0",
     "node-fetch": "^2.1.1",
-    "probot": "^9.2.5",
+    "probot": "^9.6.6",
     "queue": "^4.5.0",
     "simple-git": "1.105.0",
     "what-the-diff": "^0.4.0"
   },
   "devDependencies": {
     "@types/bunyan": "^1.8.5",
-    "@types/express": "^4.16.0",
+    "@types/express": "^4.17.2",
     "@types/fs-extra": "^5.0.1",
     "@types/jest": "^23.3.5",
-    "@types/node": "^10.12.0",
-    "@types/node-fetch": "^2.1.1",
+    "@types/node": "^10.17.9",
+    "@types/node-fetch": "^2.5.4",
     "@types/sinon": "^5.0.5",
     "jest": "^23.6.0",
-    "sinon": "^7.0.0",
+    "sinon": "^7.5.0",
     "smee-client": "^1.0.1",
     "ts-jest": "^23.10.4",
-    "tslint": "^5.11.0",
-    "tslint-config-airbnb": "^5.11.0",
-    "typescript": "^3.1.3"
+    "tslint": "^5.20.1",
+    "tslint-config-airbnb": "^5.11.2",
+    "typescript": "^3.7.3"
   },
   "engines": {
     "node": ">= 7.7.0",
